Ignore common k8s kubeconfig users, clusters, & contexts
